# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Chili Platform API
  version: 0.1.0
  summary: Serverless Django API for Chili Platform.
  description: |
    JWT-authenticated API for Chili Platform. Django runs on Cloudflare Workers
    with D1 (SQLite at the edge) and R2 for uploaded media.

    **Auth:** send `Authorization: Bearer <access>` after registering or
    exchanging credentials at `/api/auth/token/`. Access tokens last 30 minutes.
    Refresh tokens last 14 days, rotate on use, and the previous refresh token
    is blacklisted. `POST /api/auth/logout/` revokes the current refresh token.
    New accounts are unverified until they open the link in the verification
    email. Until then, forum writes, game create/update, listings, and checkout
    are rejected. Login still works, and registration still returns a JWT pair.

    **List endpoints** from the games, forum, store, and marketplace routers are
    paginated (`page`, 20 per page) unless noted. Nested post comments, marketplace
    categories, and marketplace tags are returned as a raw array.

    **Games** start as private Bitsy projects from the creator. `POST
    /api/games/{id}/release/` turns a project into a released game and keeps
    the Bitsy `data`. Release does not create a marketplace listing. Public
    lists return released games only. Bitsy `data` is returned to the owner
    and to a buyer who has that game in their library. `PATCH` accepts a PNG
...

...
    Payment is confirmed by the Stripe webhook and by
...

...

...
servers:
...
security: []
paths:
...
  /api/auth/token/refresh/:
    post:
      tags: [Auth]
      operationId: refreshToken
      summary: Refresh access token
      description: |
        Issue a new access token and a new refresh token. The refresh token
        sent in the body is blacklisted.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenRefreshRequest"
      responses:
        "200":
          description: New access token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenRefreshResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Access token from `/api/auth/token/` or `/api/auth/register/`."
    opsToken:
      type: apiKey
      in: header
      name: X-Ops-Token
      description: "Must match the Worker `OPS_TOKEN` (dev default `chili-dev-ops-token`)."
...
  responses:
...
    Unauthorized:
      description: Missing, invalid, or insufficient credentials.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorDetail"
          examples:
            jwt:
              value:
                detail: Authentication credentials were not provided.
            ops:
              value:
                detail: Unauthorized.
...
  schemas:
...
    ErrorDetail:
      type: object
      required: [detail]
      properties:
        detail:
          description: DRF may return a string or a nested error structure.
          oneOf:
            - type: string
            - type: object
              additionalProperties: true
...
    TokenRefreshRequest:
      type: object
      required: [refresh]
      properties:
        refresh:
          type: string
    TokenRefreshResponse:
      type: object
      required: [access, refresh]
      properties:
        access:
          type: string
        refresh:
          type: string
          description: Replacement refresh token. The token sent in the request is blacklisted.
...